test: refactor fin repository
- Rename `sqlite3` package to `db` because we don't plan
  to implement another fin repository than sqlite3-based one
  in the near future.
- Open sqlite3 database inside `db.New()` with exclusive lock.
  We can avoid forgetting to specify exclusive lock.
  As a result of this change, users are now required to close
  fin repository after using it.
